refactor(hooks): rename use_theme.ts to use-theme.ts
